% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curatedAdipoArray.R
\docType{package}
\name{curatedAdipoArray}
\alias{curatedAdipoArray}
\alias{curatedAdipoArray-package}
\title{\code{curatedAdipoArray} package}
\description{
A Curated Microarrays Dataset of MDI-induced Differentiated Adipocytes
(3T3-L1) Under Genetic and Pharmacological Perturbations
}
\details{
A curated dataset of Microarrays samples. The samples are MDI-induced
pre-adipocytes (3T3-L1) at different time points/stage of differentiation
under different types of genetic (knockdown/overexpression) and
pharmacological (drug treatment). The package document the data collection
and processing. In addition to the documentation, the package contains the
scripts that was used to generate the data.
}
